+++ parking-core/src/main/java/com/example/ParkingApplication.java
@@ -1,15 +1,16 @@
-package com.example.parkingscheduler;
+package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@SpringBootApplication(scanBasePackages = {"com.example.parking"})
+@EnableJpaAuditing
 @EnableScheduling
-public class ParkingSchedulerApplication {
+@SpringBootApplication
+public class ParkingApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ParkingSchedulerApplication.class, args);
+        SpringApplication.run(ParkingApplication.class, args);
     }
-
 }
