--- parking-core/src/main/java/com/example/parking/ParkingApplication.java
@@ -1,16 +1,0 @@
-package com.example.parking;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@EnableJpaAuditing
-@EnableScheduling
-@SpringBootApplication
-public class ParkingApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ParkingApplication.class, args);
-    }
-}
