+++ parking-core/src/main/java/com/example/parking/api/searchcondition/SearchConditionController.java
@@ -6,10 +6,8 @@ import com.example.parking.config.argumentresolver.MemberAuth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @Tag(name = "조회 조건 컨트롤러")
@@ -28,11 +26,11 @@ public class SearchConditionController {
         return ResponseEntity.ok(searchConditionService.findSearchCondition(memberId));
     }
 
-    @Operation(summary = "조회 조건 수정", description = "조회 조건 수정")
-    @PutMapping("/search-condition")
-    public ResponseEntity<Void> updateSearchCondition(@Parameter(hidden = true) @MemberAuth Long memberId,
-                                                      SearchConditionDto request) {
-        searchConditionService.updateSearchCondition(memberId, request);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
+//    @Operation(summary = "조회 조건 수정", description = "조회 조건 수정")
+//    @PutMapping("/search-condition")
+//    public ResponseEntity<Void> updateSearchCondition(@Parameter(hidden = true) @MemberAuth Long memberId,
+//                                                      SearchConditionDto request) {
+//        searchConditionService.updateSearchCondition(memberId, request);
+//        return ResponseEntity.status(HttpStatus.CREATED).build();
+//    }
 }
