+++ parking-core/src/main/java/com/example/parking/application/searchcondition/SearchConditionService.java
@@ -1,17 +1,10 @@
 package com.example.parking.application.searchcondition;
 
-import com.example.parking.application.SearchConditionMapper;
+
 import com.example.parking.application.searchcondition.dto.SearchConditionDto;
-import com.example.parking.domain.parking.OperationType;
-import com.example.parking.domain.parking.ParkingType;
-import com.example.parking.domain.parking.PayType;
-import com.example.parking.domain.searchcondition.FeeType;
-import com.example.parking.domain.searchcondition.Hours;
-import com.example.parking.domain.searchcondition.Priority;
 import com.example.parking.domain.searchcondition.SearchCondition;
 import com.example.parking.domain.searchcondition.SearchConditionAvailable;
 import com.example.parking.domain.searchcondition.SearchConditionRepository;
-import com.example.parking.support.Association;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -23,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class SearchConditionService {
 
     private final SearchConditionRepository searchConditionRepository;
-    private final SearchConditionMapper searchConditionMapper;
+//    private final SearchConditionMapper searchConditionMapper;
 
     public SearchConditionDto findSearchCondition(Long memberId) {
         SearchCondition searchCondition = searchConditionRepository.getByMemberId(memberId);
@@ -47,25 +40,25 @@ public class SearchConditionService {
                 .toList();
     }
 
-    @Transactional
-    public void updateSearchCondition(Long memberId, SearchConditionDto searchConditionDto) {
-        SearchCondition newSearchCondition = createSearchCondition(memberId, searchConditionDto);
-
-        searchConditionRepository.findByMemberId(memberId).ifPresentOrElse(
-                existingSearchCondition -> existingSearchCondition.update(newSearchCondition),
-                () -> searchConditionRepository.save(newSearchCondition)
-        );
-    }
-
-    private SearchCondition createSearchCondition(Long memberId, SearchConditionDto searchConditionDto) {
-        return new SearchCondition(
-                Association.from(memberId),
-                searchConditionMapper.toEnums(OperationType.class, searchConditionDto.getOperationType()),
-                searchConditionMapper.toEnums(ParkingType.class, searchConditionDto.getParkingType()),
-                searchConditionMapper.toEnums(FeeType.class, searchConditionDto.getFeeType()),
-                searchConditionMapper.toEnums(PayType.class, searchConditionDto.getPayType()),
-                searchConditionMapper.toEnum(Priority.class, searchConditionDto.getPriority()),
-                Hours.from(searchConditionDto.getHours())
-        );
-    }
+//    @Transactional
+//    public void updateSearchCondition(Long memberId, SearchConditionDto searchConditionDto) {
+//        SearchCondition newSearchCondition = createSearchCondition(memberId, searchConditionDto);
+//
+//        searchConditionRepository.findByMemberId(memberId).ifPresentOrElse(
+//                existingSearchCondition -> existingSearchCondition.update(newSearchCondition),
+//                () -> searchConditionRepository.save(newSearchCondition)
+//        );
+//    }
+//
+//    private SearchCondition createSearchCondition(Long memberId, SearchConditionDto searchConditionDto) {
+//        return new SearchCondition(
+//                Association.from(memberId),
+//                searchConditionMapper.toEnums(OperationType.class, searchConditionDto.getOperationType()),
+//                searchConditionMapper.toEnums(ParkingType.class, searchConditionDto.getParkingType()),
+//                searchConditionMapper.toEnums(FeeType.class, searchConditionDto.getFeeType()),
+//                searchConditionMapper.toEnums(PayType.class, searchConditionDto.getPayType()),
+//                searchConditionMapper.toEnum(Priority.class, searchConditionDto.getPriority()),
+//                Hours.from(searchConditionDto.getHours())
+//        );
+//    }
 }
