+++ parking-core/src/main/java/com/example/parking/config/WebMvcConfig.java
@@ -1,8 +1,6 @@
 package com.example.parking.config;
 
 import com.example.parking.config.argumentresolver.AuthArgumentResolver;
-import com.example.parking.config.argumentresolver.parking.ParkingQueryArgumentResolver;
-import com.example.parking.config.argumentresolver.parking.ParkingSearchConditionArgumentResolver;
 import com.example.parking.config.interceptor.AuthInterceptor;
 import io.swagger.v3.oas.models.PathItem;
 import java.util.List;
@@ -20,8 +18,6 @@ public class WebMvcConfig implements WebMvcConfigurer {
 
     private final AuthInterceptor authInterceptor;
     private final AuthArgumentResolver authArgumentResolver;
-    private final ParkingQueryArgumentResolver parkingQueryArgumentResolver;
-    private final ParkingSearchConditionArgumentResolver parkingSearchConditionArgumentResolver;
 
     @Value("${cors.allowedOrigins}")
     private String[] allowedOrigins;
@@ -43,8 +39,6 @@ public class WebMvcConfig implements WebMvcConfigurer {
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
         resolvers.add(authArgumentResolver);
-        resolvers.add(parkingQueryArgumentResolver);
-        resolvers.add(parkingSearchConditionArgumentResolver);
     }
 
     @Override
