+++ parking-scheduler/src/main/java/com/example/ParkingSchedulerApplication.java
@@ -1,0 +1,14 @@
+package com.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ParkingSchedulerApplication {
+
+    public static void main(String[] args) {
+        System.out.println("!!!!");
+        SpringApplication.run(ParkingSchedulerApplication.class, args);
+    }
+
+}
